Use int16_t for all object coords to allow off-screen placement in all directions.

+++ src/loop_died.c
@@ -63,8 +63,8 @@ void z_loop_died_tick(void)
         ZCircle* c = z_pool_alloc(Z_POOL_CIRCLE);
 
         if(c) {
-            int8_t x = i8(z_fix_fixtoi(z_player.x) - 1 + z_random_int8(3));
-            int8_t y = i8(z_fix_fixtoi(z_player.y) - 1 + z_random_int8(3));
+            int16_t x = i16(z_fix_fixtoi(z_player.x) - 1 + z_random_int8(3));
+            int16_t y = i16(z_fix_fixtoi(z_player.y) - 1 + z_random_int8(3));
 
             z_circle_init(c, x, y);
         }
