+++ src/obj_enemy.h
@@ -68,11 +68,11 @@ extern ZEnemyCallback z_enemy_aiTable[Z_ENEMY_NUM];
 extern ZEnemyFlyPattern z_enemy_flyTable[Z_FLY_NUM];
 extern ZEnemyCallback z_enemy_attackTable[Z_ATTACK_NUM];
 
-extern void z_enemy_init(ZEnemy* Enemy, int8_t X, int8_t Y, uint8_t TypeId, uint8_t AiState, uint8_t AiFlags);
+extern void z_enemy_init(ZEnemy* Enemy, int16_t X, int16_t Y, uint8_t TypeId, uint8_t AiState, uint8_t AiFlags);
 extern bool z_enemy_tick(ZPoolObject* Enemy);
 extern void z_enemy_draw(ZPoolObject* Enemy);
 
-extern bool z_enemy_checkCollisions(int8_t X, int8_t Y, int8_t W, int8_t H, bool AllowMultipleCollisions);
+extern bool z_enemy_checkCollisions(int16_t X, int16_t Y, int8_t W, int8_t H, bool AllowMultipleCollisions);
 
 extern void z_enemy_setFly(ZEnemy* Enemy, uint8_t FlyId);
 extern void z_enemy_setAttack(ZEnemy* Enemy, uint8_t AttackId);
