+++ src/obj_star.c
@@ -33,7 +33,7 @@
 void z_star_init(ZStar* Star)
 {
     Star->x = z_fix_itofix(
-        i8(Z_STARS_BORDER + z_random_int8(Z_WIDTH - 2 * Z_STARS_BORDER)));
+        i16(Z_STARS_BORDER + z_random_int8(Z_WIDTH - 2 * Z_STARS_BORDER)));
     Star->y = 0;
     Star->speed = u8(Z_STAR_MIN_SPEED + z_random_int16(Z_STAR_RND_SPEED));
 }
@@ -51,11 +51,11 @@ void z_star_draw(ZPoolObject* Star)
 {
     ZStar* star = (ZStar*)Star;
 
-    int8_t x = i8(z_fix_fixtoi(star->x) + z_screen_getXShake());
-    int8_t y = i8(z_fix_fixtoi(star->y) + z_screen_getYShake());
+    int16_t x = i16(z_fix_fixtoi(star->x) + z_screen_getXShake());
+    int16_t y = i16(z_fix_fixtoi(star->y) + z_screen_getYShake());
 
-    int8_t centerOffset = i8(z_fix_fixtoi(z_player.x) - Z_WIDTH / 2);
-    x = i8(x - (Z_STARS_BORDER * star->speed / Z_STAR_MAX_SPEED)
+    int16_t centerOffset = i16(z_fix_fixtoi(z_player.x) - Z_WIDTH / 2);
+    x = i16(x - (Z_STARS_BORDER * star->speed / Z_STAR_MAX_SPEED)
                     * centerOffset / (Z_WIDTH / 2));
 
     z_draw_pixel(x, y, Z_COLOR_LIGHTBLUE + (star->speed >= Z_STAR_AVG_SPEED));
